Add files via upload
In the Python programs, we compute the corresponding semidefinite programs obtaining upper bound for the optimal winning probability with constraints and loss. In the Mathematica files, there are the optimizations of strategies for the constrained BB84 monogamy-of-entanglement game using unentangled strategies, 2 qubits and 3 qubits providing their respective optimal values and strategies attaining them.